プレイヤーダメージ時にHP HUDを動かす、敵ダメージ時の処理に枠を用意

+++ Assets/Scripts/Runtime/6.Composition/InGame/Music/MusicSyncInitializer.cs
@@ -44,6 +44,6 @@ namespace KillChord.Runtime.Composition.InGame.Music
         [Tooltip("テスト用のキュー名。")]
         [SerializeField] private string _testCue;
         [Tooltip("テスト用のBPM。")]
-        [SerializeField] private int _testBpm;
+        [SerializeField] private double _testBpm;
     }
 }
